Sicherung

+++ branches/experimentalMS/webEdition/index.php
@@ -152,21 +152,35 @@ if(is_dir(WEBEDITION_PATH . 'we/cache')){
 cleanTempFiles(true);
 cleanWEZendCache();
 weNavigationCache::clean();
-we_updater::fixInconsistentTables();
+//we_updater::fixInconsistentTables();
 
 //clean Error-Log-Table
-$GLOBALS['DB_WE']->query('DELETE LOW_PRIORITY FROM ' . ERROR_LOG_TABLE . ' WHERE `Date` < DATE_SUB(NOW(), INTERVAL ' . ERROR_LOG_HOLDTIME . ' DAY)');
+if(DB_CONNECT=='msconnect'){
+	
+} else {
+	$GLOBALS['DB_WE']->query('DELETE LOW_PRIORITY FROM ' . ERROR_LOG_TABLE . ' WHERE `Date` < DATE_SUB(NOW(), INTERVAL ' . ERROR_LOG_HOLDTIME . ' DAY)');
+	
+}
 $cnt = f('SELECT COUNT(1) AS a FROM ' . ERROR_LOG_TABLE, 'a', $GLOBALS['DB_WE']);
-
 if($cnt > ERROR_LOG_MAX_ITEM_COUNT){
-	$GLOBALS['DB_WE']->query('DELETE LOW_PRIORITY FROM ' . ERROR_LOG_TABLE . ' WHERE 1 ORDER BY Date LIMIT ' . ($cnt - ERROR_LOG_MAX_ITEM_THRESH));
+	if(DB_CONNECT=='msconnect'){
+	
+	} else {
+		$GLOBALS['DB_WE']->query('DELETE LOW_PRIORITY FROM ' . ERROR_LOG_TABLE . ' WHERE 1 ORDER BY Date LIMIT ' . ($cnt - ERROR_LOG_MAX_ITEM_THRESH));
+	}
 }
 
 
-//CHECK FOR FAILED LOGIN ATTEMPTS
-$GLOBALS['DB_WE']->query('DELETE FROM ' . FAILED_LOGINS_TABLE . ' WHERE UserTable="tblUser" AND LoginDate < DATE_SUB(NOW(), INTERVAL ' . LOGIN_FAILED_HOLDTIME . ' DAY)');
 
-$count = f('SELECT COUNT(1) AS count FROM ' . FAILED_LOGINS_TABLE . ' WHERE UserTable="tblUser" AND IP="' . $GLOBALS['DB_WE']->escape($_SERVER['REMOTE_ADDR']) . '" AND LoginDate > DATE_SUB(NOW(), INTERVAL ' . intval(LOGIN_FAILED_TIME) . ' MINUTE)', 'count', $GLOBALS['DB_WE']);
+
+//CHECK FOR FAILED LOGIN ATTEMPTS
+if(DB_CONNECT=='msconnect'){
+	$count=0;
+} else {
+	$GLOBALS['DB_WE']->query('DELETE FROM ' . FAILED_LOGINS_TABLE . ' WHERE UserTable="tblUser" AND LoginDate < DATE_SUB(NOW(), INTERVAL ' . LOGIN_FAILED_HOLDTIME . ' DAY)');
+	$count = f('SELECT COUNT(1) AS count FROM ' . FAILED_LOGINS_TABLE . ' WHERE UserTable="tblUser" AND IP="' . $GLOBALS['DB_WE']->escape($_SERVER['REMOTE_ADDR']) . '" AND LoginDate > DATE_SUB(NOW(), INTERVAL ' . intval(LOGIN_FAILED_TIME) . ' MINUTE)', 'count', $GLOBALS['DB_WE']);
+
+}
 
 if($count >= LOGIN_FAILED_NR){
 	we_html_tools::htmlTop('webEdition ');
